requirements.txt
streamlit
pandas
requests
google-generativeai
pydantic
